+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v4.4
 - Throw exception if body masses are either NaN or -ve (Issue #3130)
 - Fixed issue #3176 where McKibbenActuator is not registered and can't be serialized to XML files
 - Fixed issue #3191 where CustomJoint coordinates ordering in model files affects coordinate definitions.
+- Fixed issue #3220 Memory leak running InverseKinematicsTool repeatedly and using Kinematics analysis.
 
 
 v4.3